test: create destroy file for session
