+++ tests/functional/session/create.spec.ts
@@ -1,13 +1,5 @@
-import User from '#models/user'
 import testUtils from '@adonisjs/core/services/test_utils'
 import { test } from '@japa/runner'
-
-const user = {
-  email: 'test@example.com',
-  password: 'password',
-  first_name: 'John',
-  last_name: 'Doe',
-}
 
 test.group('Session create', (group) => {
   group.each.setup(() => testUtils.db().truncate())
@@ -140,31 +132,6 @@ test.group('Session create', (group) => {
           message: 'The password field must be defined',
           rule: 'required',
           field: 'password',
-        },
-      ],
-    })
-  })
-})
-
-test.group('Session destroy', () => {
-  test('ensure user can logout', async ({ client }) => {
-    const connectedUser = await User.create(user)
-    const response = await client.delete('/logout').loginAs(connectedUser)
-
-    response.assertStatus(200)
-    response.assertBodyContains({
-      message: 'User logged out successfully',
-    })
-  })
-
-  test('ensure user cannot logout if not authenticated', async ({ client }) => {
-    const response = await client.delete('/logout')
-
-    response.assertStatus(401)
-    response.assertBodyContains({
-      errors: [
-        {
-          message: 'Unauthorized access',
         },
       ],
     })
